+++ src/view/event.js
@@ -6,7 +6,17 @@ import {
   humanizeTime,
   createElement
 } from '../utils.js';
-import {createOffersTemplate} from './offers.js';
+
+const createOffersTemplate = (offers) => {
+  const offersList = offers.map((offer) => {
+    return `<li class="event__offer">
+        <span class="event__offer-title">${offer.title}</span>
+        &plus;&euro;&nbsp;
+        <span class="event__offer-price">${offer.price}</span>
+      </li>`;
+  }).join('');
+  return offersList;
+};
 
 const createEventTemplate = (point) => {
   const type = capitalizeFirstLetter(point.type);
